requirements.txt
# Core Dependencies
python-dateutil==2.8.2
pyyaml==6.0.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Spark
pyspark==3.5.0

# GCP
google-cloud-storage==2.14.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0

# ML
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.2

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-google==10.10.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging & Monitoring
python-json-logger==2.0.7

